fix(events): move TicketAttendeeProfile type out of 'use server' file
PROVEN root cause of the publish 500 (POST /events/[id] -> 'A use server
file can only export async functions', persistent across 3 production
deployments while GET rendered fine).

events.ts is a 'use server' file; it exported a TYPE (TicketAttendeeProfile),
and scanner-client.tsx (a client component) imported it via a mixed
value+inline-type import. Next 16 passes the build but fails action
registration at runtime, 500ing every events action including publishEvent.

Move the type to types/event.ts; import it via 'import type' in both the
action file and the client. Now events.ts exports only async functions.
